+++ CHANGELOG.md
@@ -17,6 +17,7 @@ https://github.com/nwnxee/unified/compare/build8193.34...HEAD
 - Events: added skippable event `NWNX_ON_INPUT_DROP_ITEM_{BEFORE|AFTER}` which fires when a player attempts to drop an item.
 - Events: added skippable event `NWNX_ON_DECREMENT_SPELL_COUNT_{BEFORE|AFTER}` which fires when spell count (Memorized, non-memorized, or spell-like ability) decreases.
 - Events: added skippable event `NWNX_ON_DEBUG_PLAY_VISUAL_EFFECT_{BEFORE|AFTER}` which fires when the dm_visualeffect console command is used.
+- Events: added skippable event `NWNX_ON_RUN_EVENT_SCRIPT_{BEFORE|AFTER}` which fires on all object event scripts.
 
 ##### New Plugins
 - N/A
@@ -26,6 +27,7 @@ https://github.com/nwnxee/unified/compare/build8193.34...HEAD
 - Area: GetPathExists()
 - Area: {Get|Set}AreaFlags()
 - Creature: {Get|Set}SkillPointsRemainingByLevel()
+- Creature: {Get|Set}InitiativeModifier()
 - Effect: AccessorizeVisualEffect()
 - Events: SubscribeEventScriptChunk()
 - Events: UnsubscribeEventScriptChunk()